Add a policy flag to allow file namespace access to be disabled
This allows file namespace access to be turned on for the purpose of
testing, and we use this in some of the tests, but it is disabled by
default.

The intention is to remove one difference between the Chromium and
non-Chromium versions of the sandbox.

Review URL: http://codereview.chromium.org/3127033