Split trusted_thread.cc's asm code into .S files, part 1 of 2
This is the first step in extracting the i386/x86-64 assembly code in
trusted_thread.cc into separate .S files.

We add the new .S files containing semi-automatically-extracted code,
with the asm()-quotes removed and register names converted.  These
files will be made functional in the next commit.  We include the
script that performed the extraction, as well as changes to
trusted_thread.cc that helped with the extraction.

Review URL: http://codereview.chromium.org/1795012