Split trusted_thread.cc's asm code into .S files, part 2 of 2
Fix up the code in the .S files to be callable as functions.
Add these files to the makefile.

We put the arch-specific conditionals into a .S file to avoid having
to put this logic into multiple build systems (the makefile and Gyp).

Remove the original copy of the code in trusted_thread.cc.
Remove the throwaway extraction script.

Review URL: http://codereview.chromium.org/1796007