clone(): Rename arguments that were only named correctly for x86-64
The meaning of arguments 4 and 5 is swapped between i386 and x86-64.
Before, the variables were named after the x86-64 meanings.  Rename
them to the more generic "arg4" and "arg5" so that the names cannot be
misleading.  The arguments are not interpreted by the sandbox anyway.

Review URL: http://codereview.chromium.org/3369013