Test runner: Set a limit on stack size to avoid accidental OOM
A bug in the debug code can trigger the OOM killer when running
test_debugging, which tends to kill processes at random.  Set a limit
on the stack size in order to prevent this happening when running
tests.  The bug will be fixed in a later change.

BUG=http://code.google.com/p/seccompsandbox/issues/detail?id=4
TEST="make test" fails safely on an x86-64 system where gettimeofday()
  causes a system call.

Review URL: http://codereview.chromium.org/3770011