Split sandbox.cc's asm code into .S files, part 2 of 2
Update the code in the .S files with a symbol name and copyright
header, and word wrap some text.

We put the arch-specific conditionals into a .S file, as we did with
trusted_thread_asm.S.

Remove the original copy of the code in sandbox.cc.
Remove the throwaway extraction script.

BUG=none
TEST=make test all

Review URL: http://codereview.chromium.org/4243005