Add tests for glibc functions which sometimes use the x86-64 vsyscall page
test_time and test_sched_getcpu reproduce a problem on Linux 3.1
(e.g. with the current version of Ubuntu Precise), which I'll fix in a
later change.  These tests will hang, because the untrusted thread is
killed when it attempts to do a syscall.

test_gettimeofday is included for completeness.

BUG=http://code.google.com/p/seccompsandbox/issues/detail?id=17
TEST=should fail on Linux 3.1 but pass on earlier kernel versions

Review URL: http://codereview.chromium.org/8558011